Google API Location works
