+++ app/src/main/java/org/imozerov/streetartview/ui/add/AddArtObjectActivity.kt
@@ -1,42 +1,42 @@
 package org.imozerov.streetartview.ui.add
 
 import android.app.Activity
-import android.content.Context
 import android.content.Intent
 import android.location.Location
-import android.location.LocationListener
-import android.location.LocationManager
 import android.net.Uri
 import android.os.Bundle
 import android.os.SystemClock
 import android.provider.MediaStore
 import android.support.v7.app.AppCompatActivity
 import android.util.Log
-import android.widget.Toast
 import com.bumptech.glide.Glide
 import com.google.android.gms.common.ConnectionResult
 import com.google.android.gms.common.api.GoogleApiClient
 import com.google.android.gms.location.LocationServices
 import com.google.android.gms.maps.CameraUpdateFactory
+import com.google.android.gms.maps.GoogleMap
+import com.google.android.gms.maps.OnMapReadyCallback
 import com.google.android.gms.maps.SupportMapFragment
 import com.google.android.gms.maps.model.LatLng
+import com.google.android.gms.maps.model.MarkerOptions
 import io.realm.RealmList
 import kotlinx.android.synthetic.main.activity_add_art_object.*
-import kotlinx.android.synthetic.main.fragment_art_map.*
 import org.imozerov.streetartview.R
 import org.imozerov.streetartview.StreetArtViewApp
 import org.imozerov.streetartview.storage.DataSource
 import org.imozerov.streetartview.storage.model.RealmArtObject
 import org.imozerov.streetartview.storage.model.RealmAuthor
+import org.imozerov.streetartview.storage.model.RealmLocation
 import org.imozerov.streetartview.storage.model.RealmString
 import rx.subscriptions.CompositeSubscription
 import javax.inject.Inject
+
 
 /**
  * Created by sergei on 18.02.16.
  */
 class AddArtObjectActivity : AppCompatActivity(), GoogleApiClient.ConnectionCallbacks,
-        GoogleApiClient.OnConnectionFailedListener {
+        GoogleApiClient.OnConnectionFailedListener, OnMapReadyCallback {
 
     val TAG = "AddArtObjectActivity"
 
@@ -62,15 +62,9 @@ class AddArtObjectActivity : AppCompatActivity(), GoogleApiClient.ConnectionCall
 
         (application as StreetArtViewApp).appComponent.inject(this)
 
-        val mapFragment = SupportMapFragment.newInstance()
-        supportFragmentManager
-                .beginTransaction()
-                .replace(add_art_object_map.id, mapFragment, FRAGMENT_TAG)
-                .commit()
-
-        mapFragment.getMapAsync {
-            it.moveCamera(CameraUpdateFactory.newLatLngZoom(LatLng(56.3298063, 44.0095144), 14f))
-        }
+        val mapFragment = supportFragmentManager
+                .findFragmentById(R.id.add_art_object_map) as SupportMapFragment
+        mapFragment.getMapAsync(this)
 
         add_art_object_from_gallery_button.setOnClickListener(
                 { openGalleryToSetImageForArtObjectWithId () }
@@ -82,26 +76,33 @@ class AddArtObjectActivity : AppCompatActivity(), GoogleApiClient.ConnectionCall
 
         add_art_object_save_button.setOnClickListener(
                 {
-                    println("Novikov setOnClickListener")
-
                     val realmAuthor = RealmAuthor()
                     with (realmAuthor) {
                         id = SystemClock.currentThreadTimeMillis().toString()
                         name = add_art_object_author.text.toString()
-                        description = "Some author"
+                        photo = "http://photos.state.gov/libraries/media/788/images/500x500-sample.jpg"
+                    }
+
+                    val realmAuthors = RealmList<RealmAuthor>()
+                    realmAuthors.add(realmAuthor)
+
+                    val realmLocation = RealmLocation()
+                    with(realmLocation) {
+                        address = "Some address, 34"
+                        lat = lastLocation!!.latitude
+                        lng = lastLocation!!.longitude
                     }
 
                     val realmArtObject = RealmArtObject()
                     with (realmArtObject) {
-                        author = realmAuthor
+                        authors = realmAuthors
                         description = add_art_object_description.text.toString()
                         name = add_art_object_name.text.toString()
                         id = SystemClock.currentThreadTimeMillis().toString()
                         thumbPicUrl = selectedImageUrl.toString()
                         picsUrls = RealmList<RealmString>()
+                        location = realmLocation
                         println("Novikov lastLocation:" + lastLocation.toString())
-                        lat = lastLocation!!.latitude
-                        lng = lastLocation!!.longitude
                         //                        lat = 56.307872
                         //                        lng = 44.076207
                     }
@@ -110,35 +111,13 @@ class AddArtObjectActivity : AppCompatActivity(), GoogleApiClient.ConnectionCall
                 }
         )
 
-        googleApiClient = GoogleApiClient.Builder(this)
-                .addConnectionCallbacks(this)
-                .addOnConnectionFailedListener({})
-                .addApi(LocationServices.API)
-                .build()
-
-        val locationManager: LocationManager = getSystemService(Context.LOCATION_SERVICE) as LocationManager
-        locationManager.requestLocationUpdates(LocationManager.GPS_PROVIDER, 0, 0f, object : LocationListener {
-            override fun onLocationChanged(location: Location) {
-                println("Novikov onLocationChanged")
-                lastLocation = location
-
-                val sequence = "lat: " + location.latitude + "; lng: " + location.longitude
-                Toast.makeText(applicationContext, sequence, Toast.LENGTH_LONG)
-                    .show()
-            }
-
-            override fun onStatusChanged(provider: String, status: Int, extras: Bundle) {
-                println("Novikov onStatusChanged")
-            }
-
-            override fun onProviderEnabled(provider: String) {
-                println("Novikov onProviderEnabled")
-            }
-
-            override fun onProviderDisabled(provider: String) {
-                println("Novikov onProviderDisabled")
-            }
-        })
+        if (googleApiClient == null) {
+            googleApiClient = GoogleApiClient.Builder(this)
+                    .addConnectionCallbacks(this)
+                    .addOnConnectionFailedListener({})
+                    .addApi(LocationServices.API)
+                    .build()
+        }
 
     }
 
@@ -179,7 +158,9 @@ class AddArtObjectActivity : AppCompatActivity(), GoogleApiClient.ConnectionCall
 
     }
 
-    override fun onConnected(p0: Bundle?) {
+    override fun onConnected(connectionHint: Bundle?) {
+        lastLocation = LocationServices.FusedLocationApi.getLastLocation(
+                googleApiClient);
     }
 
     override fun onConnectionSuspended(p0: Int) {
@@ -188,5 +169,11 @@ class AddArtObjectActivity : AppCompatActivity(), GoogleApiClient.ConnectionCall
 
     override fun onConnectionFailed(p0: ConnectionResult) {
         throw UnsupportedOperationException()
+    }
+
+    override fun onMapReady(googleMap: GoogleMap?) {
+        val lastLatLng = LatLng(lastLocation!!.latitude, lastLocation!!.longitude)
+        googleMap?.addMarker(MarkerOptions().position(lastLatLng).title("Here you are"));
+        googleMap?.moveCamera(CameraUpdateFactory.newLatLng(lastLatLng));
     }
 }
